Static Choropleth

Static choropleth map of France.

Usage

plot_map(
  df,
  x,
  x_range = NULL,
  date = NULL,
  map = NULL,
  factor_palette = NULL,
  color_palette = NULL,
  n_breaks = 5,
  na_color = "grey",
  border_color = "white",
  border_size = 0,
  percent = FALSE,
  si_notation = TRUE,
  keywidth = 70,
  label = FALSE,
  label_column = "label",
  label_size = 3,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  legend_title = NULL
)

Arguments

df

dataframe / tibble / datatable Dataset.

x

chr Column name of variable to be plotted in the choropleth.

x_range

vector Min and max value to be used when setting the color scale of x. If NULL, the min and max observed in x will be used. Default is NULL.

date

date Date to be plotted. If NULL, the most recent available date is used. Default is NULL.

map

sf An optional simple features dataset containing GIS data to join with df if not using one of the default maps. Default is NULL.

factor_palette

vector Named list of colors (values) and label names (keys) to use for the color scale. This option is intended for categorical variables and if present plot_map() will ignore n_breaks and color_palette.

color_palette

palette Color palette to be used. If none is provided, a reversed version of 'YlOrRd' from RColorBrewer::brewer.pal() will be used.

n_breaks

int Number of desired breaks in x; to be passed to base::pretty(). Default is 5.

na_color

chr Color to use for shapes with no data. Default is grey.

border_color

chr Color of the borders. Default is white.

border_size

int Size of the borders. Default is 0 (no borders).

percent

bool Whether to format the the legend labels as percents. Default is FALSE.

si_notation

bool Whether to format the legend labels with a limited si notation (k for thousands, M for millions). Default is FALSE.

keywidth

int Width of legend color bar. Note this is the length of the whole bar, not a single segment. Default is 70.

label

bool If TRUE, id labels will appear on the map. Default is FALSE.

label_column

chr String indicating the column name to be used for labeling if label is TRUE. Default is 'label'.

label_size

int Text size of labels; ignored if label is FALSE. Default is 3.

title

chr Title.

subtitle

chr Subtitle.

caption

chr Caption.

legend_title

chr Legend Title.

Value

ggplot map.